feat: add k8s setting for container log max workers and monitor interval
- container_log_max_workers specifies the maximum number of concurrent
workers to spawn for performing the log rotate operations. Set this
count to 1 for disabling the concurrent log rotation workflows.
- container_log_monitor_interval specifies the duration at which the
container logs are monitored for performing the log rotate operation.
This defaults to 10 * time.Seconds. But can be customized to a
smaller value based on the log generation rate and the size required
to be rotated against.

Signed-off-by: Yutong Sun <yutongsu@amazon.com>